Users can now add and remove Liked Genres from their profile
connected to the Spotify api to get a list of genres. Modified APIManager  to be all class methods and use the shared instance of NSURLSession. Added corresponding tests for new methods

basic layout for Liked Genres collection view in profile screen

adding and displaying user's liked genres

performed a squash